Se agrego la nueva tabla de precios

+++ src/app/modules/operaciones/components/asignar-atencion/asignar-atencion.component.ts
@@ -3,7 +3,12 @@ import { Component, OnInit } from '@angular/core';
 import { AngularFireDatabase } from '@angular/fire/database';
 import { AngularFirestore } from '@angular/fire/firestore';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { SimpleOuterSubscriber } from 'rxjs/internal/innerSubscribe';
 import * as uuid from 'uuid';
+
+import * as XLSX from 'xlsx';
+
+type AOA = any[][];
 
 @Component({
   selector: 'app-asignar-atencion',
@@ -13,21 +18,53 @@ import * as uuid from 'uuid';
 export class AsignarAtencionComponent implements OnInit {
 
   formularioInicial: FormGroup;
+  public listaCompleta: any = [];
   public usuarios = [];
   public servicios = [];
+  public agencias = [];
+  public direcciones: any = [];
   public uidUpdate = '';
-  tiempoatencion: any;
+  tiempoLlegada: any;
+  tiempoEspera: any;
+
+  checkBoxAgencia = false
+
+  btnEnviarVisible = true;
+  btnUpdateVisible = false;
+
+
+
+  arrayModelo = {'AGENCIA':"", 'DIRECCIONES':[]}
 
   constructor(private firestore: AngularFirestore, private db: AngularFireDatabase, private _builder: FormBuilder,) {
 
     this.formularioInicial = this._builder.group({
-      codUbicacion: ['', Validators.required],
-      nombre: ['', Validators.required],
-      direccion: ['', Validators.required],
-      placa: ['', Validators.required],
-      latitud: ['', Validators.required],
-      longitud: ['', Validators.required],
-      comentario: [''],
+      numeroReferencia: ['', Validators.required],
+      motorizado: ['', Validators.required],
+      direccion: [''],
+      direccionRecojo: ['', Validators.required],
+      distrito: [''],
+      distritoRecojo: [''],
+      departamento: [''],
+      provincia: [''],
+      razonSocial: [''],
+      ruc: [''],
+      tarifa: [''],
+      telefono: [''],
+      fechaInscripcion: [''],
+      formaPago: [''],
+      nombreCompania: [''],
+      nomprom: [''],
+      nummprom: [''],
+      c_celcia: [''],
+      c_emacia: [''],
+      c_nomcont1: [''],
+      placa: [''],
+      latitud: [''],
+      longitud: [''],
+      agencia: [''],
+      agenciadireccion: [''],
+      observacion: [''],
     });
 
     this.db.list('SuizaMoto/Usuarios').valueChanges().subscribe(val => {
@@ -36,11 +73,22 @@ export class AsignarAtencionComponent implements OnInit {
         this.usuarios.push(user)
       })      
       
-      this.formularioInicial.patchValue({nombre:this.usuarios[0]['nombre']})
+      this.formularioInicial.patchValue({motorizado:this.usuarios[0]['nombre']})
+    })
+
+    this.db.list('SuizaMotoUbicaciones/DireccionesAgencias').valueChanges().subscribe(val => {
+
+      this.agencias = val;
+      this.direcciones = this.agencias[0];
+      
+
+      this.formularioInicial.patchValue({agencia: this.agencias[0]['AGENCIA'], agenciadireccion: this.agencias[0]['DIRECCIONES'][0] })
+
     })
 
     this.firestore.collection('AtencionesCurso', ref => ref.where('timestamp','>=',new Date(new Date().setHours(0,0,0,0))).orderBy('timestamp',"desc")).valueChanges().subscribe(val => {
-      this.servicios = val;
+      const ordenamiento = val.sort((a,b) => (a['timestamp'] < b['timestamp'] ? -1 : 1))
+      this.servicios = ordenamiento;
     })
 
   }
@@ -53,14 +101,30 @@ export class AsignarAtencionComponent implements OnInit {
 
       this.servicios.forEach( val => {
 
-        if(val['timestampFinRecorrido'] != "") {
-          this.tiempoatencion = (new Date(new Date(val['timestampFinRecorrido'].seconds*1000)).getTime() - new Date(new Date(val['timestampInicioRecorrido'].seconds*1000)).getTime()) / (1000*60);
-            val['tiempoTranscurrido'] = parseInt(this.tiempoatencion);
+        if(val['timestampRegistroLlegada'] != "" && val['timestampFinRecorrido'] != "") {
+
+          this.tiempoLlegada = (new Date(new Date(val['timestampRegistroLlegada'].seconds*1000)).getTime() - new Date(new Date(val['timestampInicioRecorrido'].seconds*1000)).getTime()) / (1000*60);
+          this.tiempoEspera = (new Date(new Date(val['timestampFinRecorrido'].seconds*1000)).getTime() - new Date(new Date(val['timestampRegistroLlegada'].seconds*1000)).getTime()) / (1000*60);
+          val['tiempoLlegada'] = parseInt(this.tiempoLlegada)
+          val['tiempoEspera'] = parseInt(this.tiempoEspera)
+
+        } else if(val['timestampRegistroLlegada'] != "" && val['timestampFinRecorrido'] == "") {
+
+          this.tiempoLlegada = (new Date(new Date(val['timestampRegistroLlegada'].seconds*1000)).getTime() - new Date(new Date(val['timestampInicioRecorrido'].seconds*1000)).getTime()) / (1000*60);
+          this.tiempoEspera = (tiempoActual.getTime() - new Date(new Date(val['timestampRegistroLlegada'].seconds*1000)).getTime()) / (1000*60);
+          val['tiempoLlegada'] = parseInt(this.tiempoLlegada)
+          val['tiempoEspera'] = parseInt(this.tiempoEspera)
+
+        } else if(val['timestampRegistroLlegada'] != "") {
+
+          this.tiempoLlegada = (new Date(new Date(val['timestampRegistroLlegada'].seconds*1000)).getTime() - new Date(new Date(val['timestampInicioRecorrido'].seconds*1000)).getTime()) / (1000*60);
+          val['tiempoLlegada'] = parseInt(this.tiempoLlegada)
+
         } else {
-          if(val['timestampInicioRecorrido'] != "") {
-            this.tiempoatencion = (tiempoActual.getTime() - new Date(new Date(val['timestampInicioRecorrido'].seconds*1000)).getTime()) / (1000*60);
-            val['tiempoTranscurrido'] = parseInt(this.tiempoatencion);
-          }
+
+          this.tiempoLlegada = (tiempoActual.getTime() - new Date(new Date(val['timestampInicioRecorrido'].seconds*1000)).getTime()) / (1000*60);
+          val['tiempoLlegada'] = parseInt(this.tiempoLlegada)
+
         }
 
       })
@@ -70,17 +134,29 @@ export class AsignarAtencionComponent implements OnInit {
 
   Buscar() {
 
-    this.db.list('SuizaMoto/Usuarios', ref => ref.orderByChild('nombre').equalTo(this.formularioInicial.get('nombre').value)).valueChanges().subscribe(val => {
-      if(val != null) {
-        this.formularioInicial.patchValue({placa:val[0]['placa']})
-      }
-    })
- 
-    this.db.object('SuizaMotoUbicaciones/CodReferencia/' + this.formularioInicial.get('codUbicacion').value).valueChanges().subscribe( val => {
-      if(val != null) { 
-        this.formularioInicial.patchValue({"direccion":val["Direccion"],"latitud":val["latitud"],"longitud":val["longitud"]})
-      }
-    })
+    if(this.formularioInicial.get('numeroReferencia').value != "") {
+
+      this.db.list('SuizaMoto/Usuarios', ref => ref.orderByChild('nombre').equalTo(this.formularioInicial.get('motorizado').value)).valueChanges().subscribe(val => {
+        if(val != null) {
+
+          console.log(val);
+          
+          this.formularioInicial.patchValue({placa:val[0]['placa']})
+        }
+      })
+
+      this.db.object('SuizaMotoUbicaciones/CodReferencia/' + this.formularioInicial.get('numeroReferencia').value).valueChanges().subscribe( val => {
+        if(val != null) {
+
+          this.listaCompleta = val;
+          this.listaCompleta['UbicacionGeneral'] = val['departamento'] + "/" + val["provincia"] + "/" + val["distritoRecojo"]
+
+          this.formularioInicial.patchValue(val)
+          console.log(this.formularioInicial.value);
+        }
+      })
+
+    }
 
   }
 
@@ -91,16 +167,25 @@ export class AsignarAtencionComponent implements OnInit {
     values.uid = myId;
     values.estado = 'Activo';
     values.dtInicioRecorrido = '';
+    values.dtRegistroLlegada = '';
     values.timestampInicioRecorrido = '';
+    values.timestampRegistroLlegada = '';
     values.timestampFinRecorrido = '';
+    values.habilitadorAgencia = this.checkBoxAgencia;
     values.timestamp = new Date();
     values.fechaRegistro = formatDate(new Date(),'dd-MM-yyyy','en-US')
+    values.horaRegistro = formatDate(new Date(),'HH:mm:ss','en-US')
 
-    if(values.comentario == null) {
-      values.comentario = "";
+    if(values.observacion == null) {
+      values.observacion = "";
     }
 
-    const token = this.usuarios.find(x => x.nombre == values.nombre).token;
+    if(!this.checkBoxAgencia) {
+      values.agencia = "";
+      values.agenciadireccion = "";
+    }
+
+    const token = this.usuarios.find(x => x.nombre == values.motorizado).token;
 
     if(token != undefined) {
       this.db.object("SuizaMoto/EnvioNotificaciones").set({dtEnvio:new Date().getTime(),token:token});
@@ -111,24 +196,127 @@ export class AsignarAtencionComponent implements OnInit {
     
   }
 
+  checkboxAgencia(value: any) {
+    this.checkBoxAgencia = value.target.checked
+  }
+
+  seleccionarAgencia(value: any) {
+    const listaDirecciones = this.agencias.find( valor => valor['AGENCIA'] === value.target.value )
+    this.direcciones = listaDirecciones;
+    this.formularioInicial.patchValue({agencia: value.target.value, agenciadireccion: this.direcciones["DIRECCIONES"][0] })    
+  }
+
   ActualizarAtencion(){
+
     if(this.uidUpdate != "") {
-      console.log(this.formularioInicial.value);
-      this.firestore.collection('AtencionesCurso').doc(this.uidUpdate).update(this.formularioInicial.value);
-      this.uidUpdate = "";
+
+      this.formularioInicial.value['habilitadorAgencia'] = this.checkBoxAgencia
+      
+      this.firestore.collection('AtencionesCurso').doc(this.uidUpdate).update(this.formularioInicial.value).then(() => {
+        this.btnEnviarVisible = true;
+        this.btnUpdateVisible = false;
+
+        this.formularioInicial.reset();
+        this.uidUpdate = "";
+      });
+
     }
   }
 
-  Editar(value) {
+  Editar(value) {    
     this.firestore.collection('AtencionesCurso').doc(value).valueChanges().subscribe(val => {
-      val['placa'] = null;
       this.formularioInicial.patchValue(val);
       this.uidUpdate = val['uid']
+      this.btnEnviarVisible = false;
+      this.btnUpdateVisible = true;
     })
   }
   
   Eliminar(value) {
-    this.firestore.collection('AtencionesCurso').doc(value).delete();
+    if(confirm("¿Desea eliminar el registro?")) {
+      this.firestore.collection('AtencionesCurso').doc(value).delete();
+    } else {
+      console.log("No se elimino");
+      
+    }
+    /* this.firestore.collection('AtencionesCurso').doc(value).delete(); */
+  }
+
+  uploadData(evt: any) : void { 
+
+      let workBook = null;
+      let jsonData = null;
+      const reader = new FileReader();
+      const file = evt.target.files[0];
+      reader.onload = (event) => {
+        const data = reader.result;
+        workBook = XLSX.read(data, { type: 'binary' });
+        jsonData = workBook.SheetNames.reduce((initial, name) => {
+          const sheet = workBook.Sheets[name];
+          initial[name] = XLSX.utils.sheet_to_json(sheet);
+          return initial;
+        }, {});
+        const dataString = JSON.stringify(jsonData);
+
+        const valores = jsonData['Hoja1']
+
+        console.log(valores);
+
+        /* this.db.object('SuizaMotoUbicaciones/CodReferencia').remove() */
+        valores.forEach(val => {
+          this.db.object('SuizaMotoUbicaciones/CodReferencia/'+val["numeroReferencia"]).set(val)
+        }) 
+  
+        /* 
+  
+  
+        valores.forEach(val => {
+          console.log(val);
+                  
+        }) */
+        
+  
+        /* */
+  
+  
+      }
+      reader.readAsBinaryString(file);
+    
+    
+
+
+    /* const target: DataTransfer = <DataTransfer>(evt.target);
+    if (target.files.length !== 1) throw new Error('Cannot use multiple files');
+    const reader: FileReader = new FileReader();
+    reader.onload = (e: any) => {
+
+      const bstr: string = e.target.result;
+      const wb: XLSX.WorkBook = XLSX.read(bstr, { type: 'binary' });
+
+      const wsname: string = wb.SheetNames[0];
+      const ws: XLSX.WorkSheet = wb.Sheets[wsname];
+
+      this.data = <AOA>(XLSX.utils.sheet_to_json(ws, { header: 1 }));
+
+      this.data.forEach(val => {
+
+        this.arrayModelo['DIRECCIONES'] = []
+
+        for(let i = 0; i < val.length; i++) {
+
+          if(i == 0) {
+            this.arrayModelo['AGENCIA'] = val[0]            
+          } else {
+            this.arrayModelo['DIRECCIONES'].push(val[i])
+          }
+
+        }
+
+        this.db.list("SuizaMotoUbicaciones/DireccionesAgencias").push(this.arrayModelo)     
+      })
+
+    };
+    reader.readAsBinaryString(target.files[0]); */
   }
 
 }
